Rename stdnum.iso9362 to stdnum.bic
The new name is more descriptive and easier to remember. This makes
stdnum.iso9362 a compatibility module that can be imported with the old
name but provides a deprecation warning.
