+++ docs/stdnum.bic.rst
@@ -1,0 +1,5 @@
+stdnum.bic
+==========
+
+.. automodule:: stdnum.bic
+   :members:
