--- docs/stdnum.iso9362.rst
@@ -1,5 +1,0 @@
-stdnum.iso9362
-==============
-
-.. automodule:: stdnum.iso9362
-   :members:
